pipeline/requirements.txt
# list-maker Scripts Dependencies
# Install: pip install -r requirements.txt
#
# Used by:
#   - spotify_match.py (match songs to Spotify)
#   - sync_playlist.py (sync matched songs to playlist)
#   - tal_fetch.py (fetch TAL episodes via Firecrawl)

spotipy>=2.23.0           # Spotify API client
psycopg2-binary>=2.9.0    # PostgreSQL driver (Neon)
python-dotenv>=1.0.0      # Load .env files
thefuzz>=0.22.0           # Fuzzy string matching
python-Levenshtein>=0.23.0  # 10x faster fuzzy matching
httpx>=0.27.0             # Async HTTP client (for Firecrawl)
